requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.23
aiosqlite>=0.19.0

# LangChain & LangGraph
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langgraph>=0.1.0

# Embeddings & Vector Store
sentence-transformers>=2.2.2
chromadb>=0.5.0

# Hybrid Search & Reranking
rank-bm25>=0.2.2

# ML & Data Science
scikit-learn>=1.3.2
pandas>=2.1.4
numpy>=1.26.0
joblib>=1.3.2

# MLflow
mlflow>=2.9.2

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.2
tenacity>=8.2.3
cachetools>=5.3.2

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<4.1.0
email-validator>=2.0.0

# Logging & Monitoring
structlog>=23.2.0
prometheus-client>=0.19.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
